Updated init.sh to be smarter

+++ init.sh
@@ -1,23 +1,55 @@
 #!/bin/bash
 
 # vim
-ln -s ~/dotfiles/vim ~/.vim
-ln -s ~/dotfiles/vimrc ~/.vimrc
-ln -s ~/dotfiles/gvimrc ~/.gvimrc
+if [ ! -e ~/.vim ]; then
+  ln -s ~/dotfiles/vim ~/.vim
+fi 
+if [ ! -e ~/.vimrc ]; then
+  ln -s ~/dotfiles/vimrc ~/.vimrc
+fi
+if [ ! -e ~/.gvimrc ]; then
+  ln -s ~/dotfiles/gvimrc ~/.gvimrc
+fi
 
 # git
-ln -s ~/dotfiles/gitconfig ~/.gitconfig
+if [ ! -e ~/.gitconfig ]; then
+  ln -s ~/dotfiles/gitconfig ~/.gitconfig
+fi
 
 # tmux
-ln -s ~/dotfiles/tmux.conf ~/.tmux.conf
+if [ ! -e ~/.tmux.conf ]; then
+  ln -s ~/dotfiles/tmux.conf ~/.tmux.conf
+fi
 
 # gemrc
-ln -s ~/dotfiles/gemrc ~/.gemrc
+if [ ! -e ~/.gemrc ]; then
+  ln -s ~/dotfiles/gemrc ~/.gemrc
+fi
 
-echo "source ~/dotfiles/bashrc" >> ~/.bashrc
+# bashrc
+bashgrep=$(grep "source ~/dotfiles/bashrc" ~/.bashrc)
+if [ "$bashgrep" != "source ~/dotfiles/bashrc" ]; then
+  echo "source ~/dotfiles/bashrc" >> ~/.bashrc
+fi
 
+# Submodules
 git submodule init
 git submodule update
 
-cd ~/dotfiles/vim/bundle/command-t/
-rake make
+if [ ! -e ~/dotfiles/vim/bundle/command-t/ruby/command-t/ext.so ]; then
+  if [ -e ~/.rvm/bin/rvm ]; then
+    rubyv=`type -P ruby`
+    if [ $rubyv != '/usr/bin/ruby' ]; then
+      echo
+      echo !!!!!!!!!!!!!!!!!!!!!!!!!!!
+      echo
+      echo rvm is installed, set ruby to system and rerun script!
+      echo
+      echo !!!!!!!!!!!!!!!!!!!!!!!!!!!
+      echo
+    fi
+  fi
+
+  cd ~/dotfiles/vim/bundle/command-t/
+  rake make
+fi
